requirements.txt
Adafruit-Blinka==8.64.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-motor==3.4.17
adafruit-circuitpython-motorkit==1.6.18
adafruit-circuitpython-pca9685==3.4.19
adafruit-circuitpython-register==1.10.4
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.82.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.15.0
pypubsub==4.0.0
